[react-interactions] Fix memory leak in React Flare
Tidy up logic
